3.3-permissions/api_with_restrictions/api_with_restrictions/urls.py
from django.contrib import admin
from django.urls import path, include

from rest_framework.routers import DefaultRouter

from advertisements.views import AdvertisementViewSet, FavoriteViewSet


router = DefaultRouter()
router.register('advertisements', AdvertisementViewSet, basename='advertisement')
router.register('favorites', FavoriteViewSet, basename='favourite')

urlpatterns = [
    path('api/', include(router.urls)),
    path('admin/', admin.site.urls),
]
