3.3-permissions/api_with_restrictions/requirements.txt
Django
django-filter
djangorestframework
psycopg2-binary
python-dotenv
djangorestframework-simplejwt
